% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangulation_sim.R
\name{triangulation_sim}
\alias{triangulation_sim}
\title{triangulation_sim}
\usage{
triangulation_sim(
  size = 5000,
  center = 0,
  deviation = 3,
  LeftPosition = 1,
  RightPosition = -2,
  ElectionGravity = 0.1,
  LeftWinGravity = 0.05,
  RightWinGravity = 0.05,
  randomness = 0.1,
  Shift = 0,
  TriangulationShift = 0.11,
  elections = 20,
  trials = 1
)
}
\arguments{
\item{trials}{}
}
\value{

}
\description{
triangulation_sim
}
